added better response to setup

+++ cresco.sh
@@ -5,9 +5,13 @@ case "$1" in
     docker compose -f docker/docker-compose-docs.yml up
     ;;
   setup)
-    cp scripts/pre-commit .git/hooks/pre-commit
-    chmod +x .git/hooks/pre-commit
-    echo "Pre-commit hook installed!"
+    if cmp -s scripts/pre-commit .git/hooks/pre-commit; then
+      echo "Pre-commit hook already up to date"
+    else
+      cp scripts/pre-commit .git/hooks/pre-commit
+      chmod +x .git/hooks/pre-commit
+      echo "Pre-commit hook installed/updated!"
+    fi
     ;;
   *)
     echo "Usage: $0 {serve_docs}"
